Avoid exposing implementation details
